+++ apps/sapa-dev/pages/main/potal/App.jsx
@@ -1,0 +1,7 @@
+import { start } from "@elf-framework/sapa";
+
+import { SampleApp } from "./SampleApp";
+
+start(<SampleApp />, {
+  container: document.getElementById("root"),
+});